+++ package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^6.8.0"
   },
   "devDependencies": {
+    "@types/node": "^18.14.0",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
     "@types/react-linkify": "^1.0.1",
