+++ src/App.tsx
@@ -1,9 +1,12 @@
 import { Routes, Route } from "react-router-dom";
 
-import Join from "./Join";
-import Layout from "./components/Layout";
+import Join from "@/routes/Join";
+import Layout from "@/components/Layout";
+import SecretWaitlistSkipper from "@/routes/SecretWaitlistSkipper";
+import Home from "@/routes/Home";
 
 function App() {
+  let active = localStorage.getItem('active') || "";
   return (
     <div className="App">
 
@@ -12,7 +15,8 @@ function App() {
       <Routes>
         {/* <Route path="/" element={<Home />} /> */}
         <Route path="/" element={<Layout />}>
-          <Route index element={<Join />} />
+          <Route index element={active === "true" ? <Home /> : <Join />} />
+          <Route path="secretwaitlistskipper" element={<SecretWaitlistSkipper />} />
         </Route>
       </Routes>
     </div>
