+++ src/components/WaitlistAlreadyIn.tsx
@@ -1,5 +1,6 @@
 import Notes from "./Notes";
 import { NostrProvider } from "nostr-react";
+import { useState } from "react";
 
 const relayUrls = [
     "wss://nostr.zebedee.cloud",
@@ -12,18 +13,28 @@ const relayUrls = [
 ]
 
 export function WaitlistAlreadyIn() {
+    const [skipCount, setSkipCount] = useState(0);
+
+    function skipCounter() {
+        setSkipCount(skipCount + 1);
+        if (skipCount >= 6) {
+            window.location.href = "/secretwaitlistskipper";
+        }
+    }
+
     return (
-        <main className='flex flex-col gap-4 py-8 max-w-xl mx-auto items-center drop-shadow-blue-glow'>
+        <main className='flex flex-col gap-2 sm:gap-4 py-8 px-4 max-w-xl mx-auto items-center drop-shadow-blue-glow'>
             <h1 className="text-4xl font-bold">You're on a list!</h1>
             <h2 className="text-xl">
                 We'll message you when Mutiny Wallet is ready.
             </h2>
-            <div className="p-8 rounded-xl bg-half-black">
+            <div className="px-4 sm:px-8 py-8 rounded-xl bg-half-black">
                 <h2 className="text-sm font-semibold uppercase">Recent Updates</h2>
                 <NostrProvider relayUrls={relayUrls} debug={true}>
                     <Notes />
                 </NostrProvider>
             </div>
+            <button className="opacity-0" onClick={skipCounter}>Skip</button>
         </main>
     );
 }
