+++ vite.config.ts
@@ -2,6 +2,8 @@ import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import { VitePWA } from 'vite-plugin-pwa'
 
+import * as path from 'path'
+
 // https://vitejs.dev/config/
 export default defineConfig({
 
@@ -29,4 +31,7 @@ export default defineConfig({
       enabled: true
     }
   })],
+  resolve: {
+    alias: [{ find: '@', replacement: path.resolve(__dirname, './src') }]
+  }
 })
